Migrate getPermsAsOldBitMask()
