+++ core/src/org/labkey/core/user/SecurityAccessView.java
@@ -106,7 +106,7 @@ public class SecurityAccessView extends VBox
                 continue;
 
             SecurityPolicy policy = child.getPolicy();
-            Collection<Role> allRoles = SecurityManager.getEffectiveRoles(policy,_principal);
+            Collection<Role> allRoles = SecurityManager.getEffectiveRoles(policy, _principal);
             allRoles.remove(RoleManager.getRole(NoPermissionsRole.class)); //ignore no perms
 
             List<Role> roles = new ArrayList<>();
